+++ analytics_toolkit/sql/dml/__init__.py
@@ -3,7 +3,7 @@ from .io.gp_cancel import gp_cancel_all_running_queries
 from .io.read_sql import read_sql
 from .load.load_df import build_load_df_plan, load_df
 from .table import (
-    drop_many_partitions,
+    drop_paritions,
     gp_create_many_partitions,
     gp_vacuum,
 )
@@ -12,7 +12,7 @@ from .transfer.flow.api import build_transfer_table_plan, transfer_table
 __all__ = [
     "build_load_df_plan",
     "build_transfer_table_plan",
-    "drop_many_partitions",
+    "drop_paritions",
     "execute_sql",
     "gp_create_many_partitions",
     "gp_cancel_all_running_queries",